pyproject.toml
[tool.black]
target-version = ['py310']
line-length = 120
skip-string-normalization = true
skip-magic-trailing-comma = true
force-exclude = '''
/(
  | docs
  | setup.py
)/
'''

[tool.isort]
py_version = 310
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_third_party = []
known_first_party = []
known_local_folder = []
# style: black
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
split_on_trailing_comma = true
lines_after_imports = 2
force_single_line = true
skip_glob = ["docs/*", "setup.py"]
filter_files = true

[tool.ruff]
target-version = "py310"
line-length = 160
indent-width = 4
extend-exclude = [
    "docs",
    "test",
    "tests",
    "*.ipynb"
]

[tool.ruff.lint.per-file-ignores]
"cyber/models/world/autoencoder/magvit2/models/**.py" = ["ALL"]
"cyber/models/world/autoencoder/magvit2/modules/**.py" = ["ALL"]
"cyber/models/world/autoencoder/magvit2/reconstruct.py" = ["ALL"]
"cyber/models/perception/mapping/**.py" = ["ALL"]
"cyber/models/world/dynamic/genie/**.py" = ["ALL"]

[tool.ruff.lint]
extend-ignore = ["T203"]  # Allow logging.warning
select = ["F", "E"]
extend-select = ["W", "C90", "I", "N", "B", "A", "C4", "PERF", "RUF", "T201"]
ignore = ["I001"]
fixable = ["ALL"]
unfixable = []
preview = true
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
force-single-line = true
force-sort-within-sections = false
lines-after-imports = 2
known-first-party = []
known-local-folder = []
known-third-party = []
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.mccabe]
max-complexity = 24

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
ignore-fully-untyped = true

[tool.ruff.lint.pylint]
max-args = 5
max-branches = 12
max-locals = 15
max-statements = 50

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
preview = false
docstring-code-format = true

[tool.mypy]
# Platform configuration
python_version = "3.10"
# imports related
ignore_missing_imports = true
follow_imports = "silent"
# None and Optional handling
no_implicit_optional = false
strict_optional = false
# Configuring warnings
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
warn_return_any = false
# Untyped definitions and calls
check_untyped_defs = false
disallow_untyped_calls = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_decorators = false
# Disallow dynamic typing
disallow_subclassing_any = false
disallow_any_unimported = false
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_generics = false
# Miscellaneous strictness flags
allow_untyped_globals = true
allow_redefinition = true
local_partial_types = false
implicit_reexport = true
strict_equality = true
# Configuring error messages
show_error_context = false
show_column_numbers = false
show_error_codes = true
exclude = ["docs", "test", "tests"]

[tool.poetry]
name = "cyber"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
accelerate = "0.30.1"
huggingface-hub = "0.25.2"
matplotlib = "3.9.2"
mup = "1.0.0"
omegaconf = "2.3.0"
pandas = "2.2.3"
pillow = "10.4.0"
requests = "2.32.3"
scikit-learn = "1.5.2"
scipy = "1.14.1"
setuptools = "75.1.0"
timm = "1.0.9"
tqdm = "4.66.4"
transformers = "4.41.0"
xformers = "0.0.26.post1"
lpips = "0.1.4"
ffmpeg-python = "0.2.0"
einops = "0.8.0"
decord = "0.6.0"
jxlpy = "0.9.5"
wandb = "^0.18.5"
av = "13.0.0"
imageio = {extras = ["ffmpeg"], version = "^2.36.0"}
prettytable = "^3.11.0"
types-tqdm = "^4.66.0.20240417"
opencv-python = "^4.10.0.84"
types-requests = "^2.32.0.20241016"


[tool.poetry.group.dev.dependencies]
build = ">=1.2.1"
twine = ">=5.1.0"
setuptools = ">=73.0.1"
virtualenv = ">=20.26.3"
ipython = ">=8.18.1"
ipykernel = ">=6.29.5"
ruff = ">=0.6.2"
black = ">=24.8.0"
isort = ">=5.13.2"
pre-commit = ">=4.0.0"
nbqa = ">=1.8.7"
mypy = ">=1.11.1"
mypy-extensions = ">=1.0.0"
pytest = ">=8.3.2"
pytest-cov = ">=5.0.0"
pytest-benchmark = ">=4.0.0"
codecov = ">=2.1.13"
mkdocs = ">=1.6.0"
mkdocstrings = ">=0.25.2"
mkdocs-material = ">=9.5.33"
mkdocstrings-python = ">=1.10.8"
pygments = ">=2.18.0"
types-requests = "^2.32.0.20241016"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
